Legends: Refactoring and rewrites of legend components to simplify components & reuse
