+++ packages/grafana-ui/src/components/Graph/GraphWithLegend.story.tsx
@@ -4,7 +4,7 @@ import { select, text } from '@storybook/addon-knobs';
 import { withCenteredStory } from '../../utils/storybook/withCenteredStory';
 import { GraphWithLegend, GraphWithLegendProps } from './GraphWithLegend';
 
-import { LegendPlacement, LegendDisplayMode } from '../Legend/Legend';
+import { LegendPlacement, LegendDisplayMode } from '../Legend/types';
 import { GraphSeriesXY, FieldType, ArrayVector, dateTime, FieldColorModeId } from '@grafana/data';
 
 export default {
