+++ packages/grafana-ui/src/components/GraphNG/GraphNG.story.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { withCenteredStory } from '../../utils/storybook/withCenteredStory';
 import { GraphNG } from './GraphNG';
 import { dateTime } from '@grafana/data';
-import { LegendDisplayMode } from '../Legend/Legend';
+import { LegendDisplayMode } from '../Legend/types';
 import { prepDataForStorybook } from '../../utils/storybook/data';
 import { useTheme } from '../../themes';
 import { text, select } from '@storybook/addon-knobs';
