+++ packages/grafana-ui/src/components/GraphNG/types.ts
@@ -3,7 +3,7 @@ import { DataFrameFieldIndex } from '@grafana/data';
 /**
  * Mode to describe if a legend is isolated/selected or being appended to an existing
  * series selection.
- * @public
+ * @alpha
  */
 export enum GraphNGLegendEventMode {
   ToggleSelection = 'select',
@@ -12,7 +12,7 @@ export enum GraphNGLegendEventMode {
 
 /**
  * Event being triggered when the user interact with the Graph legend.
- * @public
+ * @alpha
  */
 export interface GraphNGLegendEvent {
   fieldIndex: DataFrameFieldIndex;
